Moved the one test that is actually an integration test into an integration test dir
