+++ tox.ini
@@ -30,13 +30,13 @@ deps = -rrequirements/test.txt
 passenv =
     TEST_API_TOKEN
 commands =
-    pytest -s tests/unit --test_api_token={env:TEST_API_TOKEN:} --test_api_url={env:TEST_API_URL:}
+    pytest -s tests/unit 
 
 
 [testenv:integration]
 deps = -rrequirements/test.txt
 commands =
-    pytest -s tests/integration {posargs}
+    pytest -s tests/integration --test_api_token={env:TEST_API_TOKEN:} --test_api_url={env:TEST_API_URL:}
 
 
 [testenv:lint]
